refactor(web_core): unify ValidationConfig across validating, processing, and state layers
- Consolidate validation options into the canonical ValidationConfig in validator.ts
- Add maxDepth to TopologyOptions and forward to cycle detection in analyzeTopology
- Re-export ValidationConfig and presets from message-processor.ts to eliminate duplicate definitions
- Update SurfaceComponentsModel validateTopology and validateReferences to use ValidationConfig
- Remove TopologyValidationConfig
- Add unit tests verifying maxDepth option in SurfaceComponentsModel
